- Moves populate players job in its own directory
