+++ backend/jobs/mod.ts
@@ -1,16 +1,20 @@
 import { log } from "../deps.ts";
-import { Context, populatePlayers } from "./src/populatePlayers.ts";
+import { Context, populatePlayers } from "./src/footballApi/populatePlayers.ts";
 import {
   fetchPlayers,
   saveFootballApiPlayers,
-} from "./src/datasources/player.ts";
+} from "./src/footballApi/datasources/player.ts";
 import {
   completeJob,
   findLastJob,
   initJob,
   updateJobIdWithCurrentTeam,
-} from "./src/datasources/job.ts";
-import { fetchTeams, getTeamsIds, saveTeams } from "./src/datasources/team.ts";
+} from "./src/footballApi/datasources/job.ts";
+import {
+  fetchTeams,
+  getTeamsIds,
+  saveTeams,
+} from "./src/footballApi/datasources/team.ts";
 import { newSeason } from "./src/types.ts";
 
 const context: Context = {
