+++ backend/jobs/src/footballApi/datasources/job.ts
@@ -1,6 +1,6 @@
-import { executeQuery, ObjectId } from "../../../shared/mongoUtils.ts";
-import { Bson, Collection } from "../../../deps.ts";
-import { Job } from "../types.ts";
+import { executeQuery, ObjectId } from "../../../../shared/mongoUtils.ts";
+import { Bson, Collection } from "../../../../deps.ts";
+import { Job } from "../../types.ts";
 
 type JobCollection = Omit<Job, "id"> & { _id: ObjectId };
