+++ backend/jobs/src/footballApi/utils/apifootball/types.ts
